compressed and reduced backend fonts. adding logic to append new content to messages if timestamp has not changed
