+++ frontend/package.json
@@ -16,6 +16,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.16.0",
+    "sass": "^1.71.1",
     "use-sound": "^4.0.1",
     "websocket": "^1.0.34"
   },
