+++ frontend/src/index.html
@@ -8,6 +8,10 @@
 </head>
 <body>
 
+  <div class="background-container">
+    <div class="stars"></div>
+    <div class="twinkling"></div>
+  </div>
   <div id="root"></div>
 
 </body>
