+++ frontend/src/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
-import App from './components/App';
-import styles from './styles/global.css';
+import App from './components/App.jsx';
+import styles from './styles/global.scss';
 
 const router = createBrowserRouter([
   {
