--- frontend/src/styles/global.css
@@ -1,104 +1,0 @@
-/* Styles shared by components */
-
-/*
-Point fontfaces to server-side fonts.
-- Reduced DinaRemaster and Gohu to latin characters for smaller filesize (https://stackoverflow.com/questions/2635423/way-to-reduce-size-of-ttf-fonts
-*/
-@font-face {
-  font-family: 'DinaRemaster';
-  src: url('../../../backend/static/DinaRemasterSmaller-Regular.ttf') format('truetype');
-  font-weight: normal;
-  font-style: normal;
-}
-
-@font-face {
-  font-family: 'DinaRemaster';
-  src: url('../../../backend/static/DinaRemasterSmaller-Bold.ttf') format('truetype');
-  font-weight: bold;
-  font-style: normal;
-}
-
-@font-face {
-  font-family: 'GohuFont uni14 Nerd Font Mono';
-  src: url('../../../backend/static/GohuFontSmaller.ttf') format('truetype');
-  font-weight: normal;
-  font-style: normal;
-}
-
-@font-face {
-  font-family: 'ProggyVector';
-  src: url('../../../backend/static/ProggyVector-Regular-Mac.woff') format('woff');
-  font-weight: normal;
-  font-style: normal;
-}
-
-body {
-  position: relative;
-  background-color: rgb(34, 35, 33);
-  overflow: hidden;
-}
-
-textarea {
-  box-sizing: border-box;
-  border: 1px solid rgba(136, 136, 136, 0.41);
-  background-color: rgba(47, 43, 36, 0.982);
-  color: antiquewhite;
-  border-radius: 4px;
-  font-size: 16px;
-  font-family: "ProggyVector";
-  outline: none;
-  padding: 8px;
-  resize: none;
-}
-
-a {
-  font-style: italic;
-  color: rgba(250, 235, 215, 0.767);
-  text-decoration: none;
-}
-
-.no-select {
-  user-select: none;
-}
-
-.app-avatar {
-  height: 50px;
-  width: 50px;
-  /* background-color: rgba(226, 245, 218, 0.098); */
-}
-
-/* Width */
-::-webkit-scrollbar {
-  width: 8px; 
-}
-
-/* blend track with background */
-::-webkit-scrollbar-track {
-  background: rgb(34, 35, 33);
-}
-
-::-webkit-scrollbar-thumb {
-  background: #888;
-  border-radius: 6px;
-}
-
-::-webkit-scrollbar-thumb:hover {
-  background: #555;
-}
-
-/* unsure about using .not-connected for pre-lobby just yet */
-/* .not-connected {
-  display: grid;
-  grid-template-rows: 30px;
-  background-color: rgba(226, 245, 218, 0.775);
-  color: rgb(34, 35, 33);
-  border-radius: 10px;
-  margin-top: 15px;
-  justify-content: center;
-  text-align: left;
-  font-family: "GohuFont uni14 Nerd Font Mono";
-  font-size: 25px;
-  transition: 1s;
-  animation: wiggle 2s infinite;
-}
-*/
