+++ frontend/src/styles/lobby.scss
@@ -1,0 +1,392 @@
+/* styling for Lobby.js */
+.lobby {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.lobby-h {
+  display: flex;
+  /* flex-direction: row-reverse; */
+  align-items: center;
+  justify-content: space-between;
+  max-width: 800px;
+  width: 100%;
+}
+
+.lobby-welcome {
+  font-size: 25px;
+  font-family:"GohuFont uni14 Nerd Font Mono";
+  color: rgb(226, 245, 218);
+  text-shadow: 0px 2px 4px rgb(56, 56, 56);
+  user-select: none;
+}
+
+.user-container {
+  display: flex;
+  gap: 10px;
+  user-select: none;
+}
+
+.user-title {
+  font-family: "ProggyClean Nerd Font Mono";
+  /* background-color: rgba(250, 235, 215, 0.105); */
+  font-size: 20px;
+  font-weight: bold;
+  align-self: center;
+  border-radius: 20%;
+  padding: 2px;
+}
+
+.buttons-container-h {
+  display: flex;
+  gap: 10px;
+}
+
+.settings {
+  border: none;
+  border-radius: 4px;
+  background: none;
+  transition: .5s;
+  cursor: pointer;
+}
+
+.settings:hover {
+  transform: rotate(90deg);
+  transition: .5s;
+}
+
+.leave-lobby {
+  /* font-family: "ProggyVector"; */
+  /* font-size: 50px; */
+  border: none;
+  background: none;
+  /* color: rgba(255, 23, 7, 0.553); */
+  height: 50px;
+  width: 50px;
+  border-radius: 4px;
+  transition: .5s;
+  cursor: pointer;
+}
+
+.leave-lobby:hover {
+  transform: scale(1.08);
+  transition: .5s;
+  /* font-size: 45px; */
+}
+
+.lobby-content {
+  display: flex;
+  background-color: rgba(51, 53, 57, 0.698);
+  border-radius: 5%;
+  border-style: outset;
+  border: red;
+  flex-direction: column;
+  align-items: center;
+  max-width: 800px;
+  width: 100%;
+}
+
+.lobby-body {
+  width: 100%;
+  margin: 0 auto;
+  max-height: calc(76vh); /* needed to subtract the height of the lobby-footer to avoid overlap */
+  min-height: calc(76vh);
+  overflow-y: scroll;
+  /* border-bottom: 1px solid rgb(49, 50, 46); */
+  /* box-shadow: 0px 20px 0px rgba(49, 50, 46, 0.281); */
+  scrollbar-width: thin;
+  scrollbar-color: rgb(68, 65, 60) rgba(34, 35, 33, 0); /*thumb and track colors*/
+}
+
+.lobby-c {
+  display: flex;
+  border: 1px solid rgb(226, 245, 218);
+  border-radius: 4px;
+  background-color: rgb(34, 35, 33);
+  color: rgb(226, 245, 218);
+  text-shadow: 0px 2px 4px rgb(45, 46, 44);
+  font-family: "ProggyVector";
+}
+
+.message-list {
+  display: flex;
+  flex-direction: column-reverse;
+  gap: 10px;
+}
+
+.message-separator {
+  height: 10px;
+}
+
+.system-message {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  align-self: center;
+  color: #b5b3b0;
+  /* background-color: #b5b3b006; */
+  font-family: "DinaRemaster";
+  border-radius: 5px;
+  cursor: default;
+  user-select: none;
+}
+
+.message {
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+  // max-width: 50%;
+  overflow-wrap: break-word;
+  word-break: break-word;
+  cursor: default;
+  white-space: pre-line;
+}
+
+.message-c {
+  font-family: "ProggyVector";
+  margin: 10px;
+  line-height: 1.5;
+  // font-size: 18px;
+}
+
+/* current user msg */
+.message-cr {
+  align-self: flex-start;
+  // align-items: flex-end;
+  padding-left: 10px;
+  padding-right: 10px;
+  margin-right: 10px;
+  border-radius: 10px;
+  color: antiquewhite;
+}
+
+.message-cr:hover {
+  background-color: rgb(46, 47, 44);
+}
+
+/* other user in lobby msg */
+.message-cl {
+  // color: rgb(34, 35, 33);
+  // background-color: antiquewhite;
+  color: antiquewhite;
+  align-self: flex-start;
+  align-items: flex-start;
+  text-align: left;
+  padding: 10px;
+  margin-left: 10px;
+  border-radius: 10px;
+}
+
+.message-cl:hover {
+  background-color: rgb(202, 191, 175);
+  color: rgb(29, 30, 28);
+  transition: .3s
+}
+
+.message-info {
+  display: flex;
+  align-items: center;
+  font-family: "ProggyClean Nerd Font Mono";
+  margin-bottom: -20px;
+}
+
+.time {
+  display: flex;
+  justify-content: right;
+  padding-right: 5px;
+  font-size: 13px;
+  opacity: 0.5;
+  user-select: none;
+}
+
+.user {
+  display: flex;
+  justify-content: left;
+  margin-right: 5px;
+  padding-left: 4px;
+  padding-right: 4px;
+  font-size: 20px;
+  font-weight: bold;
+  /* background-color: rgba(250, 235, 215, 0.089); */
+  border-radius: 5px;
+  user-select: none;
+}
+
+.lobby-footer {
+  display: flex;
+  flex-direction: row;
+  /* align-items: center; */
+  position: fixed;
+  bottom: 50px;
+  resize: none;
+}
+
+.expanding-text textarea {
+  width: 100%;
+  max-height: 90px;
+  min-height: 40px;
+  min-width: 50vh;
+  max-width: 50vh;
+  align-self: center;
+  scrollbar-color: rgba(47, 43, 36, 0.982) rgba(47, 43, 36, 0.982); /* hiding within background */
+}
+
+.expanding-text textarea:focus {
+  animation: pulse-white 1s infinite;
+}
+
+@keyframes pulse-white {
+  0% {
+    outline: 2px solid transparent;
+  }
+  50% {
+    outline: 2px solid rgba(136, 136, 136, 0.41);
+  }
+  100% {
+    outline: 2px solid transparent;
+  }
+}
+
+.send {
+  border: 1px solid rgb(229, 244, 240);
+  background-color: rgb(34, 35, 33);
+  color: rgb(229, 244, 240);
+  width: 80px;
+  height: 30px;
+  border-radius: 5px;
+  margin-right: 10px;
+  margin-left: 10px;
+  transition: .2s;
+}
+
+.send:hover {
+  transform: scale(1.05);
+  transition: .2s;
+  cursor: pointer;
+}
+
+.new-messages {
+  background-color: rgba(226, 245, 218, 0.107);
+  backdrop-filter: blur(8px);
+  color: rgb(202, 191, 175);
+  border: none;
+  border-radius: 5px;
+  height: 30px;
+  position: fixed;
+  top: 90px;
+  cursor: pointer;
+  display: none;
+  font-size: 16px;
+  animation: wiggle 1s infinite;
+}
+
+.new-messages.visible {
+  display: block;
+}
+
+@keyframes wiggle {
+  0% { 
+    transform: rotate(0deg); 
+  }
+  80% {
+    transform: rotate(0deg);
+  }
+  85% {
+    transform: rotate(2deg);
+  }
+  95% {
+    transform: rotate(-2deg);
+  }
+  100% {
+    transform: rotate(0deg);
+  }
+}
+
+.settings-modal {
+  display: flex;
+  flex-direction: column;
+  position: fixed;
+  top: 30%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  border: none;
+  border: 1px solid #cccccc8e;
+  border-radius: 10%;
+  padding: 20px;
+  color: antiquewhite;
+  background-color: rgba(233, 233, 233, 0.209);
+  box-shadow: 0 0 10px rgba(0, 0, 0, 0.2);
+  z-index: 1000;
+  transition: .5s;
+}
+
+.modal-overlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  backdrop-filter: blur(4px);
+  background: rgba(0, 0, 0, 0.5);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  z-index: 999;
+  transition: backdrop-filter 1s ease;
+}
+
+.settings-header {
+  display: grid;
+  grid-template-rows: 50px auto;
+  user-select: none;
+}
+
+.settings-credits {
+  display: grid;
+  grid-template-rows: 35px;
+}
+
+.settings-close {
+  position: fixed;
+  border: none;
+  background: none;
+  top: 15px;
+  right: 10px;
+  transition: .2s;
+}
+
+.settings-close:hover {
+  transform: scale(1.2);
+  transition: 0.2s;
+  cursor: pointer;
+}
+
+/* small screen/mobile alterations */
+@media only screen and (max-width: 668px) {
+  .lobby-welcome {
+    font-size: 20px;
+    font-family:"GohuFont uni14 Nerd Font Mono";
+    color: rgb(226, 245, 218);
+    text-shadow: 0px 2px 4px rgb(56, 56, 56);
+    text-align: center;
+    user-select: none;
+  }
+
+  .lobby-content {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    width: 100%;
+  }
+
+  .lobby-footer {
+    display: flex;
+    flex-direction: row;
+    /* align-items: center; */
+    position: fixed;
+    bottom: 50px;
+    resize: none;
+    max-width: 375px;
+  }
+}
